deleteClient() flags Client objects for deletion by reallyDeleteClient()

+++ qr/Client.cpp
@@ -21,6 +21,7 @@ Client::Client(int sd, struct sockaddr_in *peer) {
 	hasInstanceKey = false;
 	game = NULL;
 	serverRegistered = false;
+	deleteMe = false;
 	char *code = NULL;
 	#ifndef _WIN32
 	if(gi != NULL)
@@ -44,6 +45,10 @@ void Client::handleLegacyIncoming(char *buff, int len) {
 	}
 }
 void Client::handleIncoming(char *buff, int len) {
+	if (deleteMe) {
+		reallyDeleteClient(this);
+		return;
+	}
 	lastPing = time(NULL);//treat anything as a ping response
 	if(len < 3) {
 		deleteClient(this);
